pyproject.toml
[project]
name = "trajan-backend"
version = "0.1.0"
description = "Trajan API - Lightweight developer workspace backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "sqlmodel>=0.0.20",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.26.0",
    "uvicorn[standard]>=0.27.0",
    "alembic>=1.13.0",
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (needed for FastAPI Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
